Rewrite jira data gatherer to more OOP
The data output is the same as before except for a timestamp that is now
correctly parsed in comments and the changelog, which is different for
various reasons, one of which is that changelog IDs are now correctly
calculated. The JIRA source data has also changed which makes it somewhat
difficult to compare and determine which version is correct, but for now
provably correct changelog data is not a major priority.

This version is configurable (arguments and config file) and the JIRA fields
are specified in a JSON file.