Track git SHA hashes manually
This makes the GitPython library not die with a fatal execution error on
a new clone (even though the documentation states that it should raise
an exception that we can handle) and allows us to (re)start a run with
an already checked-out repository. More rigid in case of intermediate
errors.

The functions in this file need to be changed to be more OOP and less
magic (right now lists and dicts are updated in a function because
python passes these by reference).