Retrieve correct date from history files
The date in the main metrics "row" dictionary is the date the quality
measurements were done, while the date in the tuple is the date since
which the metric's value is the same.

Both values may be of use, but to match with sprint data we need the
real measurement date.