Metric options: Initial version
This makes it possible to safely retrieve the metric target options from
the project definitions, to find out when (and why) a specific metric
target was changed from the default.

Requires a checkout of the quality metrics SVN repository, and the hqlib
package. Full integration in the import pipeline follows in a later
commit.