Parse project sources from definition
This adds a script and a new project definition parse, next to the
metric options parser. The project sources are to be used for
determining which repositories we can gather, rather than using
a hardcoded list in the scraper script. We may add more sources for this
information (such as the Gitlab API) if this ends up being insufficient
or extraneous.