Gitlab: Retrieve all commit comments
This allows us to retain the commit comments, usually associated with
commits on a feature branch with a pull request. Because this is
somewhat tedious within the GitLab API, use a forked version of the
python module which supports obtaining comments for a specific commit
and go through all commits using a local clone. Full Git+GitLab
integration can follow later since this is mostly meant for archival
purposes for now.