Work on streamlining project sources, credentials
Export data gathered from the project definition in project_sources.py
and read this in a Project object that can be reused by other scripts.

Convert gitlab_to_json.py so that it makes use of the project sources
that are available after the export of project_sources.py, instead of
relying on hardcoded repositories, URLs and credentials. This makes it
suitable for use in other projects.

Consolidate more credentials for projects.
Introduce the notion of subprojects.