Project sources: Domain-oriented, object-oriented
This separates the sources from the project object so that each of them
can live as its own object, with specializations for GitLab validation.
For now, these source domain objects live separately from the Subversion
and Git implementations, but the goal is to make them implement these
source interfaces for better integration.