Project sources: Improve source persistence
Make Source objects hashable by their exported data, which makes it
easier to check if a Source object exists and to manage adding and
removing them.

Use the Project source management in project_sources.py, and also in
gitlab_to_json.py to register the additional repositories in the project
sources.