Version control: Generalize Git multi-repo holder
Support source repositories more generally and agnostic, so do not
distinguish between Git and Subversion.
Reduce internal state of the Git repository object so that it matches
more with the Subversion repository. This is also a step toward fully
integrating domain sources and repository objects.
Change the output field names of versions and the file name.