Version control: Lazy-load repo, fix diff stats
This replaces the current repository back-end solution where the
repository is local at first, but replaced by a remote repo when
initialized from a URL. This did not work for Subversion repos without
a local checkout. Instead lazy-load the repo, which gives us time to
alter the repo to use.

Also fix diff stats parsing so that we can retrieve these statistics
from the repositories. Make the sprint data optional (with warning).