Jira: Retrieve all versions from Jira
Even if a version is not referenced by some issue, then we still want
information about that version. Also, the start date of a version is not
contained in the issue result. Retrieve all versions, even archived ones
from the API before searching issues so that all fields are populated.