Jira: Support more operations on Tables
Tables can now be searched for row existence and its values updated.
For key-based tables, the search uses only the keys, while the update
only allows changes to fields that are not part of a unique identifier.

This also clean up member variable access. This is groundwork for using
these operations for updating issue links with more data along the way.