Jira: Clean up fields and table sources
Fields, special fields and type cast parsers can now specify the name of
the table they wish to register. Previously, this had to come from the
field specification.
Consolidate fields and special field parsers such that they share
a parent class that has the `parse` method to ensure it exists, and also
provides the search_field property so that we can register special
fields in the exact same way as normal fields.

Initial work for augmenting issue link data: track the pieces of text
that define inward/outward links.