Jira: Changelog field base class interface
Add a base class for changelog fields and change the method
specification of parse_changelog to allow the fields to use more
information from the issue resource. When creating fields for the
changelog entries, support using existing fields if they meet the base
specification.

This is groundwork for collecting start and end dates for issue links.