Jira: Retrieve start and end dates of issue links
Use changelog data to keep track of when links were added and deleted.
For simplicity, we only allow one time range per issue link pair,
consisting of (from, to, relationship, direction). The inward/outward
direction was added for proper matching of the links, which are all
bidirectional in nature.

To do this, heavy use is made of the new Table updates, the prefetcher,
as well as some internal storage for matching changelog strings with
their associated relationship type. Also, the issue links now use the
keys of the issue, and not their IDs. This requires upgrading the
importer as well.