Git source: SSH key-based authentication environment
This makes the Git repository retrieval support SSH as a communications
protocol. This also adds support for reading SSH keys from a credentials
file pointed at by an environment variable, such that the private
identity file is not stored in the repository.

This support full integration with the 'secrets file' functionality of
the credentials binding plugin of Jenkins. Store a private id_rsa SSh
identity file without a passphrase in the credentials store and add
a 'secrets file' credentials variable to do this.

The SSH URL is automatically generated from the public HTTP URL, which
was obtained from the project definition, when environmental credentials
are used. This works for GitLab URLs and probably other repository
managers as well.