Version control: Intregate with Source
Replace `from_url` with `from_source` which also provides the repo name.
This allows us to use credentials settings from within the repository
object, such as credential paths and whether to trust unsafe hosts.
This also forms a basis for moving GitLab processing into a class, so
that the repository commit iteration process does not need to be
duplicated and can be customized with credentials settings.