Gitlab: Support reading legacy dropin files
This makes it possible to load files made with gitlab_to_json.py, even
those that were renamed to include the former group name. The data is
loaded through the GitLab API to ensure data validity, and replaces the
usual API calls (thus only performing version information retrieval).

Update the source objects to allow access to more 'original' source
information, including the plain URL, path names of other URLs than the
final one, and the original namespace in the GitLab URL. Describe the
process of replacing the path group component better.

Pass along the Project domain object to version control repository
objects so that we can find the export path. The project may also be
helpful for other uses later on, but for now only GitLab uses it.