Gitlab: Retrieve dropins directly from directory
The dropins that are usually exported by a script are copied to the
export directory. Legacy dropins are not. Retrieve those directly from
the dropins directory.

Also update the scraper such that it can be executed from within the
repository, as well as with the repository checked out at scripts/
according to the Jenkins build setup.