Git: Support source-based revision limiting
Add settings to credentials that can specify a different revision range
using tags or dates. 'tag' limits the end revision to the commit with
that tag, while 'from_date' limits the starting point of commits from
the master branch from migration purposes. The two are not used together
and 'tag' received priority if the tag exists in the repo.