Requirements: Remove regex from dependencies
Do not require regex to run the scraper. Only the topdesk gatherer
requires this library, and this data can be imported manually.
Jenkins instances are unable to install this dependency if they do not
have a `gcc` installed and no wheels are available for the platform.

Add a README to specify installation instructions and overview of the
repository.