Version control: Provide Source object
Update callers that directly use the constructor, but were still passing
it a repository name instead of a Source object. Some of these contexts
do not have URLs, but the Source object should handle that no problem.
Build the Source object of the project definitions repository using
settings, which are now actually project-wide using a meta-class.