Git: Retrieve tags, fix local timestamps
Extract the tags that the Git repository has. Since there is no simple
interface for extracting the tags of individual commits, we just grab
all of them at the end, similar to GitLab merge requests.

Keep track of the auxiliary data in actual Table objects, which is no
longer JIRA-specific. This makes sure that no duplicate data is added in
the exported JSON.

Add a new utility function for converting to local datetimes so that the
GitLab legacy commit matcher works correctly with dates from different
timezones (must be converted to 'local' epochs).