Version control: Clean up stats and auxiliary data
Add a new method `get_data` that indicates that we want auxiliary data,
instead of abusing `get_versions` for this. Also do not depend on the
value of `retrieve_stats`, since one may want diff stats but not the
auxiliary data or vice versa.

In fact, remove `retrieve_stats` as a member variable/property and more
it to keyword arguments. This allows us to store the `project` argument
of the constructor instead, and pass along our criteria on method call
time and perform the auxiliary data retrieval with additional arguments.
This also removes the need for changing `has_commit_comments` in the
GitLab repository, instead we determine this based on an additional
parameter as well.

Also streamline Subversion and Git behavior to only consider trunk and
master. This speeds up the Subversion version retrieval and is a first
step to handling tags (and branches) separately in Svn.