Version control: Retrieve tags from Subversion
Subversion repositories with a tags directory in their root should have
only directories containing the frozen state of the trunk at a certain
version which is a (candidate) release. These directories should not be
changed after they have been created. These assumptions allow us to
query the Subversion repository easily for the existing tags and output
in the same way as the Git tags, except for missing a message (which is
usually automated anyway).