Utils: Support Python 3-based unicode encoding
Many data gatherers use the `parse_unicode` function to verify and clean
up free-text fields to ensure that the Unicode characters are formatted
correctly in JSON output. In Python 3, unicode and str are the same type
and conversion yields bytes objects. Streamline both versions and types
by passing through encode and decode, thus outputting the same data in
python 2 and 3.