Docker: Speed up incremental builds
Perform the installation of dependencies before the copy step of the
data gathering scripts. This makes docker builders with memory make use
of the cache of the partial build, since the requirements will most
likely not have changed. This does mean that version enforcements in
requirements.txt are necessary to force an update of these dependencies.

This reduces build times of subsequent builds from four and a half
minutes to about ten seconds.