File store: Initial interface and dropins POC
Support retrieving and putting directories and files at an external file
store, which can be implemented by several backends. Implementation of
ownCloud backend is provided. This is helpful for storing (large) sets
of dropin files for (archived) projects. These are then no longer stored
in the repository, but instead are retrieved at the stage that we need
them.