Update trackers: Support different external source types
Convert to a more object-oriented approach for better reuse of methods
and compatibility with later use within controller API.

Support both direct MonetDB database access as well as a new SSH
connection, which is meant to be a gateway between the agents and the
database, taking care of some synchronization tasks as well.

Add a new parameter 'files' to restrict retrieving certain files; for
SSH this is required, for DB source this acts as an optional filter.