Controller/agent: Distribute SSH keys from agent
Instead of letting the controller generate the public/private key pair,
do so in the agent and distribute the public keys to the controller.
Optionally distribute the SSH public key to a GitLab host as well.

Upgrade the controller to expect a public key on POST instead of
providing an attachment download. This also paves the way for an API
response body. Clean up the public key update process.

Add HTTPS SSL certificate verification support.