Controller API: Offload actions to a daemon
Use Pyro4 to connect with a named daemon and communicate with it with
a local networked socket. This adds security because the CGI process no
longer has to have increased privileges, and the daemon is not exposed
to other networks and thus less likely to be compromised.

This requires some setup with Pyro daemons, including a name server,
which is more smooth with some systemd service files. Documentation on
this is currently kept externally.