Version control: Prior check whether repository is up to date
Allow version control repository objects to check whether a source is up
to date before it is even retrieved. This is used by GitLab repositories
to check whether the HEAD commit reported by the API is the same as the
latest collected commit. This lets us skip the repository without clone
or pull commands to check the commit state afterward, thus saving time
in incremental updates of old repositories, especially if repo cleanup
is enabled in the scraper.