GitLab sources: Do not clone repositories, cleanup
The repositories were cloned in advance for the git_to_json.py script
and to determine whether a repository is empty. Because the
git_to_json.py script might not need a clone to operate on GitLab
repositories (e.g. when there are no updates), this is wasteful. Instead
check whethere there are any commits through the API for the latter
check, and skip the entire repository retrieval.

Also clean up now unused log ratio argument.