Docker: More automated scraping with cron/Jenkins
Support running the agent in an automated fashion, either as
a standalone docker container, within a BigBoat dashboard or as
a Jenkins job. Support registering a periodic cron job. Generate a key
at startup and perform the scraping in the job. Setup the agent user
correctly. The docker job only retrieves VCS commits, which can then be
exported to a controller server.