Version control: Retrieve change path type
Determine if the change is a modification, additiona, deletion, or
replace (Git only, move/copy) of a file. This allows us to find when
files are changed in this way more robustly, and to know when files are
recreated. This is based on both diffs of a merge if possible.

Also fix file paths in Git such that they do not contain suffix/prefix
split paths and old => new paths in case of a move/copy.