Git change paths: Ignore upstream merged files
Merge commits have multiple parents (usually two), of which one is the
head of the branch being merged and the other is the previous head of
the branch merged into. We try both to find out about changed files, but
not all files have statistics deduced for them. Now that we can simulate
the change path names provided by the stats, this provides a good
indication of which paths are 'changed' in the merge commit, be it from
the merge or conflict resolution.