Quality metrics: Parse default targets
Use the quality reporting hqlib repository to extract previous versions
and the default values of (low/perfect) values defined in the metrics.
This can be used to find out which metrics have their values changed,
and to compare it with metric targets in the project definitions.

Old files are Python 2 only while new files can only be parsed with
Python 3. We use an AST parser to obtain the target values.