Version control: Extract authored date
Git has a separation of commit date and author date, where the authored
date provides an indication of when work was started on the commit.
Projects that make use of amends, squashes or rebases will have such
dates that differ from the commit date, which would be the moment of
amendment, squash or rebase.