Git: Add support for TFS API
This adds support for retrieving data from the Team Foundation Server
API, which includes push events, pull requests (with reviewer votes) and
other repository information for extra sources.

This uses NTLM authentication and a separate API connection object for
specific project collections, as indicated by the path in the source.