Git: Change path diff stats performance
Improve the performance of the statistics retrieval for the change path
differences. The git python library does not cache the commit.stats
property, thus retrieving it multiple times results in many calls to
processes that obtain difference stats (git diff --numstat). This
command can be slow for large diffs, thus resulting in a nonlinear
increase in complexity. Use a local variable to speed this up.