TFS: Extract project from remote URL
For TFS 2015 and up, the project name may be placed before the _apis
endpoint part, or for some requests between the API area part and the
remaining resource route parts. For TFS 2013, the project name, if it is
provided, must be after the area part (for the supported API endpoints).
This complicates the URL parsing since projects may appear in the front
in the remote URL, even though this may not always work for the API URL.
They at least do for the repositories query results from the API.