SVN: Remove Jenkins checkout into script
This adds functionality to the version control repository objects to
checkout or update the repository from remote state. For Git, this was
already mostly available, just not as a separate method.

Use the SVN update/checkout functionality to retrieve the quality
metrics definitions repository using a script. This removes the need to
use Bash commands to check for the local state, change to the correct
directory and perform the commands, using the URL. The repository source
data now uses the settings.

Update the svn library to make use of the latest features. Pin it so
that we can check the compatibility before upgrading.