GitLab: Read normal dropin files
The Jenkins scraper already handles dropins for full scripts, but for
Git auxiliary data not all export/update files may have dropins.
Specifically, we want to find out if we need to perform additional
requests on an archived repository, based on the existence of these
dropin files. This is mostly relevant for the commit comments file.

This also clean up the dropins parser hierarchy to be able to use the
parser for other purposes.