Controller: Separate update trackers from export
The controller server now creates two subdirectory structures in the
agent user's directory: export and update. The latter is now filled with
the current state of the update trackers for the project, instead of the
export directory. The agent itself may still place both in the export
directory (locally and remotely). This seperation ensures that the agent
will continue to retrieve all necessities, even if the controller has
not updated the database yet. This ensures we have no loss of data from
agents that track their own state rather than the current one.