JIRA: Extract 'environment' field
Some projects use this field to describe something similar to an epic
(theme of issues), a specifier for the development phase the issue
occurs in (OTAP model), or the component that is affected by the issue
or receives the feature (sometimes a version). It is mostly freeform
text but has some structure in some projects.