History to JSON: Use formatter URLs/paths
This makes the path and URL more versatile in situations where the path
or URL does not end with the quality metrics name, but it is placed
somewhere in the URL, for example. Project-specific overrides are still
supported. Environment-specific configuration may need updates to
include a formatter argument after the path.