History to JSON: Support GitLab downloads/clones
Depending on whether --export-url or both --export-url --export-path are
provided (the latter is done in the scraper), history_to_json.py now
understands GitLab repository URLs and respectively alters the URL to
a raw download of the history file, or (sparsely) clones the repository
similar to retrieve_metrics_repository.py does. This makes the script
support authenticated and unauthenticated access as well as preliminary
cloning and on-line downloads during the import.