GitHub: Initial API support
Allow sources to have GitHub personal access tokens which provide
limited access to the GitHub API. This adds support for discovery of
related sources from the environment of a source repository, defined by
the repos owned by a user or organization. Optionally the environment
can be further demarcated by a GitHub team that manages the repos via
a custom argument to the source.

A specialized repository object is also created but does not perform any
additional tasks on its own.