GitHub: Unstub repository further
Actually register the repository class in the Source object, and make
the main API repo result more readily available since most other calls
relevant to the repository build upon it.

Make the repository a review system-based repo and define which tables
and (API) calls we are going to call, with the parsing of the repo data
itself as first realization.