GitHub: Limit API to new issue (comments)
The issues API seems to work well with limiting based on a 'since'
parameter, formatted as ISO 8601 with T-delimiter and Zulu time
(although in practice the API is not so strict in its parsing), so be
sure to provide it with a UTC time (instead of local tracker time).

This reduces the number of requests to the API and download size when
a repository with many issues has been partially imported. The parameter
does not function/exist for pull request and pull request review
(commit) comments, and for pull request issue comments the python
library does not support the argument at this moment (or we need to hack
around to make the pull request think it is an issue). The local update
checks are still necessary for safety.