Docker: Register agent key in project definitions GitLab
The generated SSH key must also be distributed to the GitLab for the
project definitions, as well as registering a credentials section for
the host. Otherwise, the agents are unable to authenticate against
a protected GitLab host for the project definitions and they fail to
retrieve the definitions. Document the new variables.