Exporter: Use Jenkins API
This replaces makeshift code for interfacing with the Jenkins build API
with the jenkins module with a nicer interface. Provide getters to
create job, view or job build objects based on their name easily.

Remove the 'crumb' setting for jenkins since the CSRF protection
endpoint is now always checked upon initialization.

Move the exporter daemon to the main directory such that it can reach
gatherer scripts and configuration. It made little sense to have
a directory with only one file while other exporter scripts live in the
main directory (such as controller-export.sh).