Git: Implement remote up-to-date check
Check whether the repository is up to date before cloning it or pulling
in changes. This can reduce the time and disk space spent on
repositories that are already imported. It was already implemented for
GitLab through the API but now works for all remote Git repositories,
including TFS, GitHub and SSH/'dumb' remotes.

The internal command used, git ls-remote, specifically does not need
a cloned repository as long as the remote is provided as the same URL
used to clone. This and the environment credentials can be obtained from
the source information. We limit the ls-remote command to only check the
master head, as in other uses in the Git collectors.