GitHub: Determine type using last diff hunk line
The line type of a pull request comment was previously deduced using the
position or original position. There are cases where the diff hunk does
not match up with the line numbers. In cases where they do match up, the
diff hunk is trimmed to the position we want. Thus simply retrieve the
last line from the hunk to determine the line type of the comment.