Source: Support custom ports
Allow customizing the port, independently from the host change (although
it respects host changes, and then checks for port changes), and support
SSH URLs with port numbers in them. Ensure domain name does not start
with a dash for security. This is helpful when an SSH server to connect
to a source (e.g. Git) is not running on the standard port.