Exporter: Run import when no import tasks are busy
Emulate Jenkins behavior of keeping a job in the queue if another job
with the same name is running. Check whether any other Jenkins scrapers
are being run by the exporter, and wait for them to finish. This makes
concurrency issues with the database less likely and reduces sudden I/O
load when multiple agents signal that they are done at the same time.
Concurrency may still arise with the Jenkins jobs that are separately
spawned, but the database should be able to handle this.