Git: Support lowercase shared repository config
The Git clone command writes configuration in a canonical way, i.e., all
variable names are lowercased. This is not our expectation and with the
limitation of GitPython not adhering the case-insensitivity of the Git
config file, we should just try both the documented name and the
lowercase name for the sharedRepository configuration.