BigBoat: New export format and database import
The BigBoat status data now contains numeric values and the maximum
values, if possible for each status item.

The controller API now instructs the gatherer daemon to import the
status information into the database. If this fails (database is not
ready or project is not yet known), then the data is written to
a controller file for later inspection and import.