History to JSON: Override to legacy file
This is useful for sources that have switched to compact history, but we
explicitly want to read the legacy file. This requires the command line
argument, if given, to override the data source property, and to have
the option to enable or disable it. This situation also requires
overriding the filename.