History to JSON: Configurable compression
The metrics do not define any compression, but the storage system
(GitLab or Jenkins) may provide it in a compressed format due to the
quality reporting scripts that are used to create them. Thus we need
a setting and an argument to define this for all our possible sources.
This also allows disabling a predefined compression, and adds
extensibility for other compression formats than just GZip.