Source: Export environments
After collection/expansion of the project/environment sources, we export
them for later re-use. However, the sources also provide useful metadata
about their source locations, but the Java import cannot access that
anymore. Thus we export this data to JSON, which includes for each
environment its identifier (string/tuple) and the type of the
representative source.