History to JSON: Add option to delete repository
This deletes the repository before performing a shallow clone, ensuring
that the repository tracks the most recent version instead of not
updating locally. This is a problem in Git before 1.9, when shallow
fetches were made more powerful to actually track file changes instead
of just the commits. It is recommended to enable this to ensure the
shallow clone is always up to date before using it.