GitLab: Only use API date to fail up-to-date check
Do not depend on the API-provided last activity date to consider the
date to be up to date, because some GitLab instances are incorrectly
configured with regard to time zones and/or respond with a UTC time
which is actually a local time. Furthermore, some instances had a broken
activity indicator which would make this code consider them to be always
up to date. It is safe to return that the data is out of date, though.