JIRA: Remove empty fields from output, project/key
Instead of outputting the zero string '0' for fields that have no value,
simply make the value `None` and let the field output handle skipping
this field. The importer must be made aware of the potential of missing
fields. This saves a lot of space in the output with ever-growing fields
for different types of issues, optional fields that only some projects
use, and changes which rarely happen. This speeds up the gathering and
importing steps for JIRA.

Not all fields are suitable for skipping, and sometimes the zero string
means something else (amount of 0, text content '0'). These fields are
now separable from the default zero string. This change only affects the
main issue data table, not other tables from JIRA or elsewhere.