Controller: Create hush login file
This writes a file in the agent's home directory that disables the
message showing the last login date when attempting an interactive
login. If the agent logs in to a restricted shell, then this file cannot
be deleted, so the login date is not made known.

Also replace use of 'os.mknod', which does not use its arguments
correctly, with a file-write plus permission/ownership calls.