Controller: Skip replacing home directory if keys match
The agent's home directory may contain data from a previous instance,
but if the public key matches with the registered one, then this means
it is simply a generate_key call at the start of a scrape. Avoid slowing
down the scrape by doing useless operations.

Also, normalize newlines of the generated key so that it can match with
the controller key and GitLab key.