Controller: Check remote version as well
If the controller server is outdated but the agent is up to date with
the remote repository, then still consider the agent to be up to date.
Add more sanity checks of the remaining parts of the version so that it
aligns with the gatherer version and the default branch.