Environment sources: Replace sources with same URL
This makes sure that a source from the project definition that has the
same URL as another source gets replaced with a source from the
environment, using the name as stated in the environment instead of the
one stated in the project definition. This is relevant when multiple
sources are used within the same domain object in the project
definition, such that they can obtain different names from the
environment.