Controller: Make public key readable before verifying
This speeds up the agent registration process if the public key is the
same. Because authorized_keys files must be owner-readable only on many
platforms, we need to broaden the permissions before we can perform the
equality check on existing keys.

Document that the permissions must always be corrected afterward.