Source: Per-protocol credentials username
Allow credentials sections to have 'username.ssh' and 'username.http'
keys which override the username for the specific protocol. If one of
these keys exists then the 'username' key is not used for this protocol.
Other functionality remains the same, so if the found key has a falsy
value then it will fall back to the URL's username.