Version control: Skip up to date check if not pulling
If we are not going to update the local state of the repository, then it
is not necessary to check if there are new versions. In this case we
always want to collect the local revisions (fetched by other means) and
the auxiliary data, thus the check and potential skip is not performed.