Project definition: Mock Sonar to avoid requests, stuffing
The current quality report classes cause the Sonar class to perform
requests upon initialization, which may not be able to reach their
destination but use the default 15-second timeout. Additionally, the
class is replaced and has a different class name according to detected
version number. Avoid this for the source ID parsing.