Sonar to JSON: Collect pages, start/end measures
Fix the way we collect the first and last measures so that we do not
have missing measures. Also keep on collecting if the API returns pages
of results. Write the collection date of each metric to an update file
so that we skip collecting the old values of the same metric over again.