Gitlab: Strip path from web URL
If a strip path is provided for the host, then remove the strip path
from the URL before applying more alterations to it. This makes sure
that the URL can be converted to an SSH one, and that the GitLab group
or namespace can be extracted.

This makes it possible to convert URLs with strip paths from web to SSH
or the other way around.