Docker: Upgrade Alpine version tag
This should allow installing a version of Rust which is needed for some
cryptography installations. Another option is to upgrade the Python
version in the tag as well, since newer versions support binary Python
package distributions with the musllinux target. But that might bring
some other changes along, so we try a newer Alpine first.