Consider Python 3.10 as maximum version for now
Some outdated requirements that use lxml at a fixed version mean the
latest Python headers are incompatible (moved in Python 3.11), so unless
we work around the old versions like we do with python-gitlab, we are
stuck with that version. Further testing would be necessary to ensure
this is possible, but for now use Python 3.10.