GitHub Actions workflow: mypy typing coverage, for combined Sonar gate
Our unit tests only cover the gatherer module, not the scripts and
daemons that are also in this repository. Still, we wish to have some
metric of code quality for these (so that Sonar's New Code gate does
have something to measure and does not just deny new script/daemon code)
and typing at least gives an indication of code integrity. Coveralls
still provides the unit test coverage measurements, so it seems fine to
use SonarQube for this (a private instance only had typing coverage from
analysis run by Jenkins).